Keep the twin predicate off Rigor::Environment's pinned public surface
`public_api_drift_spec.rb` pins `Rigor::Environment`'s singleton surface
as the ADR-2 API plugin authors depend on, and the first cut grew it
with `bundled_overlay_twin_signatures?`. That is not what the method is:
it is an internal seam with exactly two callers inside the engine, so
adding it would have promised plugin authors something nobody meant to
promise. `private_class_method` is not available either -- one caller is
`CheckRules`, in another class.

It moves to `Rigor::Environment::RbsLoader` instead, beside
`under_gem_overlay_root?`, which answers the same question about the
OTHER copy of the same declarations and is already the documented seam
outside callers use. That class also already owns the overlay's layout
(`GEM_OVERLAY_SIGS_ROOT`, `gem_overlay_sig_paths`), and it is not a
pinned surface. `Environment` keeps the ADR-72 policy -- which gem maps
to which plugin id, and when an overlay is eligible at all -- and now
passes the id to the loader, which answers only the filesystem half. The
snapshot therefore needs no edit: the pinned surface is byte-identical
to master's again.

The two helpers behind it are `private` inside `class << self`, placed
last so the modifier cannot capture a method meant to be reachable.
`private_class_method` raises NameError there -- inside `class << self`
these are instance methods of the singleton class, so it looks for them
one level further out.

`Plugin::Loader.bundled_plugin_sig_path`, the branch's other new class
method, is checked and left public: its class is not pinned, its caller
is in another class, and it matches the visibility of the
`bundled_plugin_path` sibling it was modelled on.
