+++ docs/internal-spec/inference-engine.md
@@ -715,7 +715,7 @@ The `0.1.x` cycle added these read-side accessors (all seeded through `Environme
 - `Rigor::Environment#project_patched_methods` — the `Rigor::Inference::ProjectPatchedMethods` registry populated from the project's `.rigor.yml` `pre_eval:` list ([ADR-17](../adr/17-monkey-patch-pre-evaluation.md)). Consulted by the dispatcher's project-patched-methods tier; defaults to `ProjectPatchedMethods::EMPTY`. The registry's read surface is `#lookup(class_name:, method_name:, kind:)` and `#empty?`.
 - `Rigor::Environment#dependency_source_index` — the opt-in gem-source catalog ([ADR-10](../adr/10-dependency-source-inference.md)), consulted by the dependency-source inference tier.
 - `Rigor::Environment#synthetic_method_index` — the ADR-16 / ADR-18 substrate emit table, consulted by the synthetic-method tier.
-- `Rigor::Environment#plugin_registry` — the loaded `Rigor::Plugin::Registry` for the run (also the source of plugin-contributed `signature_paths:` merged into the RBS loader, per [ADR-25](../adr/25-plugin-contributed-rbs.md), and `open_receivers:` / `protocol_contracts:` consulted by `CheckRules` and the binder). `open_receivers:` is NOT the only source `CheckRules#unbounded_receiver_surface?` consults for the "unsound to flag" verdict: `Rigor::Analysis::CheckRules::GEM_OVERLAY_OPEN_RECEIVERS` (issue #632, tracked further by #660) is a second, plugin-independent list for a class Rigor's own bundled gem-overlay RBS (`data/gem_overlay/`, ADR-72) declares knowing the declaration is partial — an auto-applied overlay has no plugin manifest for an `open_receivers:` entry to live on. Membership there is gated on `Rigor::Environment::RbsLoader#signature_paths` actually including, this run, a directory carrying one of Rigor's own partial declarations of the gem — either the auto-applied overlay (`RbsLoader.under_gem_overlay_root?`) or the bundled plugin twin's own `sig/` (`Rigor::Environment.bundled_overlay_twin_signatures?`, the route a project takes when it wires that `sig/` through `signature_paths:` instead of `plugins:`; issue #672). Either arm satisfies the gate, so the exemption is active exactly when the RBS it protects is (ADR-26 WD1), not merely because some project somewhere owns a class of the same name.
+- `Rigor::Environment#plugin_registry` — the loaded `Rigor::Plugin::Registry` for the run (also the source of plugin-contributed `signature_paths:` merged into the RBS loader, per [ADR-25](../adr/25-plugin-contributed-rbs.md), and `open_receivers:` / `protocol_contracts:` consulted by `CheckRules` and the binder). `open_receivers:` is NOT the only source `CheckRules#unbounded_receiver_surface?` consults for the "unsound to flag" verdict: `Rigor::Analysis::CheckRules::GEM_OVERLAY_OPEN_RECEIVERS` (issue #632, tracked further by #660) is a second, plugin-independent list for a class Rigor's own bundled gem-overlay RBS (`data/gem_overlay/`, ADR-72) declares knowing the declaration is partial — an auto-applied overlay has no plugin manifest for an `open_receivers:` entry to live on. Membership there is gated on `Rigor::Environment::RbsLoader#signature_paths` actually including, this run, a directory carrying one of Rigor's own partial declarations of the gem — either the auto-applied overlay (`RbsLoader.under_gem_overlay_root?`) or the bundled plugin twin's own `sig/` (`RbsLoader.gem_overlay_twin_signatures_loaded?`, the route a project takes when it wires that `sig/` through `signature_paths:` instead of `plugins:`; issue #672). Either arm satisfies the gate, so the exemption is active exactly when the RBS it protects is (ADR-26 WD1), not merely because some project somewhere owns a class of the same name.
 - `Rigor::Environment#hkt_registry` — the lightweight-HKT type-function registry ([ADR-20](../adr/20-lightweight-hkt.md)), consulted by the HKT-builtin-return dispatcher tier. It MUST be the merge, **in this order** (last-write-wins): the bundled `Builtins::HktBuiltins.registry` base, then the plugin overlay (`Plugin::Registry#hkt_overlay_registry`, which flat-maps every loaded plugin's `hkt_registrations:` / `hkt_definitions:` in registration order), then any user `.rbs` overlay on top. The getter is memoised and MUST return `HktRegistry::EMPTY` when nothing contributes, so a registry-free run skips the merge.
 
 Slice 6 introduces fact-store access. The methods added in later slices MUST NOT change the Slice 1 surface.
