+++ docs/internal-spec/plugin.md
@@ -584,7 +584,7 @@ a plugin declaring none of them is a plain per-file analyzer:
 | `consumes` | `Array<Consumption>` | Cross-plugin facts this plugin reads (`{ plugin_id:, name:, optional: }`); drives the loader's topological ordering (ADR-9). |
 | `signature_paths` | `Array<String>` | RBS signature directories the plugin contributes, relative to the plugin gem root; resolved by `Loader` and merged into the environment (ADR-25). |
 | `owns_receivers` | `Array<String>` | Receiver class names this plugin owns for dispatch routing. |
-| `open_receivers` | `Array<String>` | Receiver class names whose method surface is unbounded — e.g. `ActiveRecord::Relation`, which delegates every user-declared `scope` to its model (ADR-26). Such a class is exempted from `call.undefined-method` outright, and from the two signature-reading rules (`call.wrong-arity`, `call.argument-type-mismatch`) for any method name it INHERITED rather than declared: a delegated name that collides with an ancestor's (`relation.open` → `Kernel#open`) resolves to a signature the call does not run. A method the open class declares itself keeps both checks; a method it genuinely has through an ancestor (`Enumerable` on a Relation) loses them too — the test is the definition site, not delegation. This field only takes effect once a project loads the declaring plugin. Two OTHER, plugin-independent sources feed the same `CheckRules#unbounded_receiver_surface?` gate: `RbsLoader#synthesized_type_names` stub types, and (issue #632, tracked further by #660) `CheckRules::GEM_OVERLAY_OPEN_RECEIVERS` — a plain constant, not a manifest field, for a class Rigor's own BUNDLED gem-overlay RBS (`data/gem_overlay/`, ADR-72) declares knowing the declaration is partial. An auto-applied overlay has no plugin manifest for an `open_receivers:` entry to live on, so a receiver it declares open (`ActiveSupport::Duration`, which forwards undeclared members to the wrapped numeric via `method_missing`) needs protection independent of whether any plugin is loaded at all; `rigor-activesupport-core-ext` ALSO lists the same class under its own `open_receivers:`, so the two sources overlap there by design — either alone is sufficient, and the constant is what covers the overlay-only case the manifest field structurally cannot reach. Membership in that constant is NOT itself sufficient, though: `CheckRules#gem_overlay_loaded?` additionally requires one of Rigor's own partial declarations of that gem to have actually loaded THIS run — `RbsLoader#signature_paths` including a directory under `RbsLoader.under_gem_overlay_root?` (the auto-applied overlay) or one matching `Rigor::Environment.bundled_overlay_twin_signatures?` (the bundled plugin twin's own `sig/`, reached when a project wires it through `signature_paths:` instead of `plugins:` — issue #672, where the overlay now stands down and the twin's manifest-less declaration is the only one loaded) — otherwise a project that never locks the gem, and happens to own a class of the same qualified name itself, would silently lose `call.undefined-method` coverage on it. This restores the same "the protection is active exactly when the RBS is" property `open_receivers:` gets for free from requiring the declaring plugin to be loaded (ADR-26 WD1). |
+| `open_receivers` | `Array<String>` | Receiver class names whose method surface is unbounded — e.g. `ActiveRecord::Relation`, which delegates every user-declared `scope` to its model (ADR-26). Such a class is exempted from `call.undefined-method` outright, and from the two signature-reading rules (`call.wrong-arity`, `call.argument-type-mismatch`) for any method name it INHERITED rather than declared: a delegated name that collides with an ancestor's (`relation.open` → `Kernel#open`) resolves to a signature the call does not run. A method the open class declares itself keeps both checks; a method it genuinely has through an ancestor (`Enumerable` on a Relation) loses them too — the test is the definition site, not delegation. This field only takes effect once a project loads the declaring plugin. Two OTHER, plugin-independent sources feed the same `CheckRules#unbounded_receiver_surface?` gate: `RbsLoader#synthesized_type_names` stub types, and (issue #632, tracked further by #660) `CheckRules::GEM_OVERLAY_OPEN_RECEIVERS` — a plain constant, not a manifest field, for a class Rigor's own BUNDLED gem-overlay RBS (`data/gem_overlay/`, ADR-72) declares knowing the declaration is partial. An auto-applied overlay has no plugin manifest for an `open_receivers:` entry to live on, so a receiver it declares open (`ActiveSupport::Duration`, which forwards undeclared members to the wrapped numeric via `method_missing`) needs protection independent of whether any plugin is loaded at all; `rigor-activesupport-core-ext` ALSO lists the same class under its own `open_receivers:`, so the two sources overlap there by design — either alone is sufficient, and the constant is what covers the overlay-only case the manifest field structurally cannot reach. Membership in that constant is NOT itself sufficient, though: `CheckRules#gem_overlay_loaded?` additionally requires one of Rigor's own partial declarations of that gem to have actually loaded THIS run — `RbsLoader#signature_paths` including a directory under `RbsLoader.under_gem_overlay_root?` (the auto-applied overlay) or one matching `RbsLoader.gem_overlay_twin_signatures_loaded?` (the bundled plugin twin's own `sig/`, reached when a project wires it through `signature_paths:` instead of `plugins:` — issue #672, where the overlay now stands down and the twin's manifest-less declaration is the only one loaded) — otherwise a project that never locks the gem, and happens to own a class of the same qualified name itself, would silently lose `call.undefined-method` coverage on it. This restores the same "the protection is active exactly when the RBS is" property `open_receivers:` gets for free from requiring the declaring plugin to be loaded (ADR-26 WD1). |
 | `type_node_resolvers` | `Array` | `Plugin::TypeNodeResolver` entries contributing custom RBS type-name resolution (ADR-13). |
 | `protocol_contracts` | `Array<ProtocolContract>` | Path-scoped behavioural contracts (`path_glob` + `method_name` + `singleton` + param/return types + severity); provide-and-check (ADR-28). |
 | `source_rbs_synthesizer` | `#call(path) -> String?` | A callable that synthesises RBS from a project source file at env-build time (e.g. rbs-inline ingestion) (ADR-32). |
